5478: Add Leaf Query Filter IT
